moved ADP

+++ src/main/java/org/tindalos/principle/domain/expectations/IntThresholder.java
@@ -2,5 +2,5 @@ package org.tindalos.principle.domain.expectations;
 
 public interface IntThresholder {
 
-    double violationThreshold();
+    int violationThreshold();
 }
